Add test to verify if arbBlockHash on pending block returns latest
