+++ system_tests/rpc_test.go
@@ -7,10 +7,8 @@ import (
 	"context"
 	"path/filepath"
 	"testing"
-	"time"
 
 	"github.com/ethereum/go-ethereum/ethclient"
-	"github.com/offchainlabs/nitro/solgen/go/mocksgen"
 )
 
 func TestIpcRpc(t *testing.T) {
@@ -25,25 +23,5 @@ func TestIpcRpc(t *testing.T) {
 	defer cleanup()
 
 	_, err := ethclient.Dial(ipcPath)
-	Require(t, err)
-}
-
-func TestPendingBlockTimeAndNumberAdvance(t *testing.T) {
-	t.Parallel()
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	builder := NewNodeBuilder(ctx).DefaultConfig(t, true)
-	cleanup := builder.Build(t)
-	defer cleanup()
-
-	auth := builder.L2Info.GetDefaultTransactOpts("Faucet", ctx)
-
-	_, _, testTimeAndNr, err := mocksgen.DeployPendingBlkTimeAndNrAdvanceCheck(&auth, builder.L2.Client)
-	Require(t, err)
-
-	time.Sleep(1 * time.Second)
-
-	_, err = testTimeAndNr.IsAdvancing(&auth)
 	Require(t, err)
 }
